refactor: extract apply_join_level_filter spine across both scans
Both `JoinScan::build_relnode_df` and `AggregateScan::build_relnode_df`
ended their `RelNode::Filter` arm with the same translate-and-filter
sequence: call `PredicateTranslator::translate_join_level_expr`,
ok_or_else into a DataFusionError, apply the filter to the DataFrame.
JoinScan additionally dropped the synthetic `mark` column post-filter
when the predicate was `MarkOrNull`.

Hoist the spine into `apply_join_level_filter` in
`customscan/datafusion/translator.rs`. The MarkOrNull cleanup moves
behind a `handle_mark: bool` parameter (true for JoinScan, false for
AggregateScan, since aggregate-on-join never produces MarkOrNull
predicates). The signature takes the inputs each caller already
gathered (translated_exprs, ctid_map, predicates list, deferred
positions, sources) so the helper can be called as a single
expression on both sides.

JoinScan's caller shrinks from ~50 LOC to ~12 LOC. AggregateScan's
shrinks from ~17 LOC to ~10 LOC. The translate-and-filter logic now
lives in exactly one place, and a future correctness fix to it (e.g.
better error formatting, extra invariants) only needs to land once.
