+++ pg_search/src/postgres/customscan/datafusion/translator.rs
@@ -21,7 +21,7 @@ use datafusion::logical_expr::{col, lit, BinaryExpr, Expr, Operator};
 use datafusion::prelude::DataFrame;
 use pgrx::{pg_sys, PgList};
 
-use crate::api::HashMap;
+use crate::api::{HashMap, HashSet};
 use crate::postgres::customscan::joinscan::build::{
     JoinLevelExpr, JoinLevelSearchPredicate, JoinNode, JoinSource, JoinType as PgJoinType,
     RelationAlias,
@@ -349,6 +349,63 @@ impl<'a> PredicateTranslator<'a> {
             _ => None,
         }
     }
+}
+
+/// Translate a `JoinLevelExpr` predicate into a DataFusion filter expression
+/// and apply it to `df`. This is the shared spine of the `RelNode::Filter`
+/// arm in JoinScan's and AggregateScan's `build_relnode_df` recursions —
+/// both modules previously inlined the same translate-and-filter sequence.
+///
+/// `handle_mark` controls JoinScan-specific cleanup: when `true`, a
+/// `MarkOrNull` predicate triggers a post-filter projection that drops the
+/// synthetic `mark` column from the schema. AggregateScan never produces
+/// `MarkOrNull` predicates and passes `false`.
+#[allow(clippy::too_many_arguments)]
+pub fn apply_join_level_filter(
+    mut df: DataFrame,
+    predicate: &JoinLevelExpr,
+    translated_exprs: &[Expr],
+    ctid_map: &HashMap<pg_sys::Index, Expr>,
+    join_level_predicates: &[JoinLevelSearchPredicate],
+    deferred_positions: &HashSet<usize>,
+    sources: &[&JoinSource],
+    handle_mark: bool,
+) -> Result<DataFrame> {
+    let filter_expr = unsafe {
+        PredicateTranslator::translate_join_level_expr(
+            predicate,
+            translated_exprs,
+            ctid_map,
+            join_level_predicates,
+            deferred_positions,
+            sources,
+        )
+    }
+    .ok_or_else(|| {
+        DataFusionError::Internal(format!(
+            "Failed to translate join level expression tree: {:?}",
+            predicate
+        ))
+    })?;
+
+    df = df.filter(filter_expr)?;
+
+    // For MarkOrNull filters in JoinScan, drop the synthetic "mark" column
+    // post-filter so it doesn't leak into the projection.
+    if handle_mark && matches!(predicate, JoinLevelExpr::MarkOrNull { .. }) {
+        let schema = df.schema().clone();
+        let proj_cols: Vec<Expr> = schema
+            .columns()
+            .into_iter()
+            .filter(|c| c.name != "mark")
+            .map(col)
+            .collect();
+        if !proj_cols.is_empty() {
+            df = df.select(proj_cols)?;
+        }
+    }
+
+    Ok(df)
 }
 
 /// Creates a DataFusion column expression with a bare table reference.
